+++ sdk/core/ts-http-runtime/src/auth/tokenCredential.ts
@@ -1,6 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+import { OAuth2Flow, OAuth2FlowType } from "./authFlows.js";
+
 /**
  * Represents a credential capable of providing an authentication token.
  */
@@ -30,91 +32,6 @@ export interface GetTokenOptions {
    * OAuth2 flow metadata used during token acquisition.
    */
   authFlows?: OAuth2Flow[];
-}
-
-/**
- * Represents the different types of OAuth2 authentication flows supported.
- * Can be either AuthorizationCodeFlow or ClientCredentialsFlow.
- */
-export type OAuth2Flow = AuthorizationCodeFlow | ClientCredentialsFlow | BaseOAuth2Flow;
-
-/**
- * Describes the OAuth2 flow type
- */
-export enum OAuth2FlowType {
-  /**
-   * Authorization code flow
-   */
-  authorizationCode = "authorizationCode",
-
-  /**
-   * Client credential flow
-   */
-  clientCredentials = "clientCredentials",
-}
-
-/**
- * Base interface for OAuth2 flow configurations.
- * Contains common properties shared across different OAuth2 flow types.
- */
-export interface BaseOAuth2Flow {
-  /**
-   * The type of OAuth2 flow. Used to differentiate between different authentication flows.
-   */
-  type: string;
-
-  /**
-   * List of scopes for the credential
-   */
-  scopes?: string[];
-}
-
-/**
- * Configuration for OAuth2 Authorization Code flow.
- * This flow is typically used in scenarios where user interaction is possible,
- * such as web applications where users can log in through a browser.
- */
-export interface AuthorizationCodeFlow extends BaseOAuth2Flow {
-  /**
-   * Specifies this is an Authorization Code flow
-   */
-  type: OAuth2FlowType.authorizationCode;
-
-  /**
-   * The authorization URL
-   */
-  authorizationUrl: string;
-
-  /**
-   * The token URL
-   */
-  tokenUrl: string;
-
-  /**
-   * The refresh URL
-   */
-  refreshUrl?: string;
-}
-
-/**
- * Configuration for OAuth2 Client Credentials flow.
- * This flow is used in machine-to-machine authentication scenarios where no user interaction is required.
- */
-export interface ClientCredentialsFlow extends BaseOAuth2Flow {
-  /**
-   * Client credentials flow
-   */
-  type: OAuth2FlowType.clientCredentials;
-
-  /**
-   * The token URL
-   */
-  tokenUrl: string;
-
-  /**
-   * The refresh URL
-   */
-  refreshUrl?: string;
 }
 
 /**
