+++ sdk/core/ts-http-runtime/src/interfaces.ts
@@ -1,8 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { OAuth2Flow } from "./auth/tokenCredential.js";
-
 /**
  * A HttpHeaders collection represented as a simple JSON object.
  */
@@ -233,16 +231,6 @@ export interface PipelineRequest {
 
   /** Settings for configuring TLS authentication */
   tlsSettings?: TlsSettings;
-
-  /**
-   * List of OAuth2 flows to be used for authentication.
-   */
-  authFlows?: OAuth2Flow[];
-  
-  /**
-   * List of scopes to be used for authentication.
-   */
-  scopes?: string[];
 }
 
 /**
